tests/wpt/tests/magnetometer/Magnetometer-supported-by-feature-policy.html

<!DOCTYPE html>
<title>Test that magnetometer is advertised in the feature list</title>
<link rel="help" href="https://w3c.github.io/webappsec-feature-policy/#dom-featurepolicy-features">
<link rel="help" href="https://w3c.github.io/sensors/#feature-policy-api">
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script>
test(() => {
    assert_in_array('magnetometer', document.featurePolicy.features());
}, 'document.featurePolicy.features should advertise magnetometer.');
</script>
